model_governance_framework/requirements.txt
evidently>=0.4
mlflow>=2.10
great-expectations>=0.18
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.31
jinja2>=3.1
plotly>=5.18
